clothes-detection/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-aarch64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=51_gnu
ailia=1.5.0.0=pypi_0
bidict=0.21.2=pyhd3eb1b0_0
blas=1.0=openblas
blinker=1.6.2=py310hd43f75c_0
bottleneck=1.4.2=py310hf6ef57e_0
brotli=1.0.9=h998d150_8
brotli-bin=1.0.9=h998d150_8
bzip2=1.0.8=h998d150_6
c-ares=1.19.1=h998d150_0
ca-certificates=2024.11.26=hd43f75c_0
cairo=1.16.0=h537eab0_5
click=8.1.7=py310hd43f75c_0
contourpy=1.3.1=py310hb8fdbf2_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h647bc0d_1
dbus=1.13.18=h821dc26_0
eigen=3.4.0=hb8fdbf2_0
expat=2.6.3=h419075a_0
ffmpeg=4.2.2=hdfaaa67_0
flask=3.0.3=py310hd43f75c_0
flask-socketio=5.3.1=py310hd43f75c_0
fontconfig=2.14.1=h881afe3_3
fonttools=4.51.0=py310h998d150_0
freetype=2.12.1=h6df46f4_0
glib=2.78.4=h419075a_0
glib-tools=2.78.4=h419075a_0
gmp=6.2.1=h22f4aa5_3
gnutls=3.6.15=hc6589d6_0
graphite2=1.3.14=h22f4aa5_1
gst-plugins-base=1.14.1=h419075a_1
gstreamer=1.14.1=h998d150_1
harfbuzz=4.3.0=hfddee83_2
hdf5=1.12.1=h2117f30_3
icu=73.1=h419075a_0
imutils=0.5.4=pypi_0
itsdangerous=2.2.0=py310hd43f75c_0
jinja2=3.1.4=py310hd43f75c_1
jpeg=9e=h998d150_3
kiwisolver=1.4.4=py310h419075a_0
krb5=1.20.1=h2e2fba8_1
lame=3.100=hfd63f10_0
lcms2=2.12=h5246980_0
ld_impl_linux-aarch64=2.40=h48e3ba3_0
lerc=3.0=h22f4aa5_0
libabseil=20240116.2=cxx17_h419075a_0
libbrotlicommon=1.0.9=h998d150_8
libbrotlidec=1.0.9=h998d150_8
libbrotlienc=1.0.9=h998d150_8
libclang=14.0.6=default_hd3a980f_1
libclang13=14.0.6=default_h5e70c7c_1
libcups=2.4.2=hb788212_1
libcurl=8.9.1=hfa2bbb0_0
libdeflate=1.17=h998d150_1
libedit=3.1.20230828=h998d150_0
libev=4.33=hfd63f10_1
libffi=3.4.4=h419075a_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h6e398d7_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hd439bcf_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h998d150_3
libidn2=2.3.4=h998d150_0
libllvm14=14.0.6=hbfe7563_4
libnghttp2=1.57.0=hb788212_0
libopenblas=0.3.21=hc2e42e2_0
libopus=1.3.1=h998d150_1
libpng=1.6.39=h998d150_0
libpq=17.2=h6ac735f_0
libprotobuf=4.25.3=h94b7715_0
libssh2=1.11.1=hfa2bbb0_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h998d150_0
libtiff=4.5.1=h419075a_0
libunistring=0.9.10=h2f4d8fa_0
libuuid=1.41.5=h998d150_0
libvpx=1.8.2=h7c1a80f_0
libwebp-base=1.3.2=h998d150_1
libxcb=1.15=h2f4d8fa_0
libxkbcommon=1.0.1=h4b4cad1_2
libxml2=2.13.5=h6097fa9_0
lz4-c=1.9.4=h419075a_1
markupsafe=2.1.3=py310h998d150_0
matplotlib=3.9.2=py310hd43f75c_1
matplotlib-base=3.9.2=py310hff4f700_1
mysql=8.4.0=h505ef0a_1
mysql-connector-python=8.4.0=py310h2e46eea_1
ncurses=6.4=h419075a_0
nettle=3.7.3=h82288b7_1
numexpr=2.10.1=py310hd351863_0
numpy=1.26.4=py310he45c16d_0
numpy-base=1.26.4=py310h15d264d_0
opencv=4.10.0=py310h5f425dd_0
openh264=1.8.0=h22f4aa5_0
openjpeg=2.5.2=ha2c532c_0
openldap=2.6.4=h418a7c9_0
openssl=3.0.15=h998d150_0
packaging=24.1=py310hd43f75c_0
pandas=2.2.3=py310h419075a_0
pcre2=10.42=hcfaa891_1
pigpio=1.78=pypi_0
pillow=11.0.0=py310h199e702_0
pip=24.2=py310hd43f75c_0
pixman=0.40.0=h2f4d8fa_1
protobuf=4.25.3=py310hce23d57_0
pyparsing=3.2.0=py310hd43f75c_0
python=3.10.15=h4bb2201_1
python-dateutil=2.9.0post0=py310hd43f75c_2
python-engineio=4.1.0=pyhd3eb1b0_0
python-socketio=5.3.0=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py310hd43f75c_0
qt-main=5.15.2=h1e70531_11
readline=8.2=h998d150_0
seaborn=0.13.2=py310hd43f75c_0
setuptools=75.1.0=py310hd43f75c_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h998d150_0
tk=8.6.14=h987d8db_0
tornado=6.4.1=py310h998d150_0
tzdata=2024b=h04d1e81_0
unicodedata2=15.1.0=py310h998d150_0
werkzeug=3.0.6=py310hd43f75c_0
wheel=0.44.0=py310hd43f75c_0
x264=1!152.20180806=h2f4d8fa_0
xz=5.4.6=h998d150_1
zlib=1.2.13=h998d150_1
zstd=1.5.6=h6a09583_0
